### 2023-07-19 #5664 @editorialbot
[REVIEW]: dcTensor: An R package for discrete matrix/tensor decomposition
**Submitting author:** @kokitsuyuzaki (<a href="http://orcid.org/0000-0003-3797-2148">Koki Tsuyuzaki</a>)
**Repository:** https://github.com/rikenbit/dcTensor
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.3
