### 2023-07-20 #5664 @dekuenstle
The submission presents a highly specialized R package for discrete matrix decomposition
algorithms.
The package is openly available (MIT-Licence on Github) and distributed according to R-standards on CRAN;
the API is simple, and extensive unit tests are available (while it is unclear if they run automatically).
However, I have some suggestions for improvement concerning the documentation and the manuscript.

The CRAN documentation provides a good explanation of the input parameters and expected data format
as well as references to publications describing the implemented algorithms.
Here it would be helpful to have a small summary of the algorithm in the Description field of each
function, i.e. describing for what purpose the algorithm is used or how it is different from the others.
The only installation instructions on the repository readme are aimed more at developers than end users because they install from Github instead of CRAN.
The readme could be a good place to give users and developers separate instructions and add the missing community guidelines.
The file LICENSE.md contains an MIT license statement; however, there is another file, LICENSE, with just a header
that might be removed.
Unit tests are available in the repository, but test coverage and automation are not documented.
The Github action workflow "build_test_push.yml",  running on push and pull, apparently is not
running the test suite–––Test automation should be added or, if already done, documented.

The paper Summary provides a high-level introduction to discrete matrix decomposition. The
arguments could profit from a few more explanations. In the first paragraph, it might be inclusive
to introduce matrix decomposition, i.e., explain that the product of factor matrices reconstruct
the original matrix and why this is a non-trivial problem. In the third paragraph,
it remains to be seen why there is a demand for tensor decomposition.
The Statement of Need justifies the need through the non-existence of algorithm implementations.
However, this justification needs more explanation for which problems to solve
 and for which target community these implementations are needed. Especially
 it would be relevant to explain the need in applications for algorithms provided in the submitted
 package but not in related works. This also provides the opportunity to integrate the very short
 Related Work section into the Statement of Need.
While I appreciate the provided code example in the manuscript, it requires further explanations
and a more coherent integration in the other sections. The plots should use identical colors and
frequency axis to compare original and reconstructed data visually.
Many citations and references do not follow the standards, are duplicates ("Zhang, Z. et al. in 2007 (Z. et al. Zhang, 2007)"),
are missing Co-authors ("CICHOCK, 2009" should be "Cichock et al., 2009"), or are permuted ("Z. et al. Zhang, 2007");
most of those could be resolved by adding the full list of authors to the Bib file.
Overall, the language in the manuscript should be revised with a focus on rephrasing rough nested sentences
and increasing coherence.
