### 2023-07-22 #5664 @CeciliaCoelho
## TL;DR
I recommend major revision of the paper and documentation available at the repository.

### Overall review
There is evidence of the effort put in the package development however the documentation of the repository is a big downside due to being very small and poor (readme) or scattered in multiple files (example the contribution instructions). Not having instructions of how to use it or information on the examples included in the package in the readme can make users reluctant in trying this package due to the extra time it takes to search for all information they need to use it. This may sound harsh but I think such a nice package should have an equally nice and friendly readme. This will attract more users.
Also, the paper is also lacking information and the summary information does not belong to the summary but rather to the statement of need.
I advise the authors to check accepted papers and their repositories for a guideline.

### Major comments
1 - The summary is too big in comparison to the rest of the documentation. The summary should describe concisely the package that is going to be presented stating what functions are implemented, some key features it contains and the target programming language. Right now the summary presented is more appropriate for the statement of need since it gives a background explanation of the implemented algorithms. The summary should be one paragraph. If the authors need some guidelines to fix the summary they should check accepted JOSS papers.
2 - The statement of need is very incomplete. It is missing a background explanation on the algorithms, what they are for and their applications. Having only a list of names with some references is very limited and at least a paragraph explaining each item should be included.
3 - The repository readme is very poor. It should have a brief explanation at the top stating what the repository has. It can be the summary used in the paper after fixing according to 2. It should clearly state how to use the functions provided and how to run the example code. The readme should also include the contributing.md. All information should be readily available to the users. Again I suggest the authors to check readme files of accepted papers if they need a guideline.
4 - The references need to be fully redone. The names of all the authors need to be written. Only when used in the text "et al." is used. It appears CICHOCK should be Cichocki.
5 - The Example section in the paper should be explained with more detail. Like what is the goal, the description of the problem
6 - The plots in Figure 1 are not consistent and cannot be used for comparison between them since they show different scales

### Minor comments
7 - There are 2 license files. Please remove the extra one
8 - In the paper the authors use "I" and "We". Just one should be used in all paper
