### 2023-08-04 rikenbit/dcTensor#8 @kokitsuyuzaki
paper.md
@dekuenstle 's comment

Summary
> The paper Summary provides a high-level introduction to discrete matrix decomposition.

> The arguments could profit from a few more explanations.

> In the first paragraph, introducing matrix decomposition might be inclusive, i.e., explaining that the product of factor matrices reconstruct the original matrix and why this is a non-trivial problem.

> In the third paragraph, it remains to be seen why there is a demand for tensor decomposition.

Statement of Need

> The Statement of Need justifies the need through the non-existence of algorithm implementations.

> However, this justification needs more explanation for which problems to solve and for which target community these implementations are needed.

> Especially it would be relevant to explain the need in applications for algorithms provided in the submitted package but not in related works.

Related Work
> This also provides the opportunity to integrate the very short Related Work section into the Statement of Need.

Misc

> Overall, the language in the manuscript should be revised with a focus on rephrasing nested sentences and increasing coherence.

@CeciliaCoelho 's comment

Summary

> The summary is too big in comparison to the rest of the documentation.

> The summary should describe concisely the package that is going to be presented stating what functions are implemented, some key features it contains and the target programming language.

> Right now the summary presented is more appropriate for the statement of need since it gives a background explanation of the implemented algorithms.

> The summary should be one paragraph. If the authors need some guidelines to fix the summary they should check accepted JOSS papers.

Statement of need

>  The statement of need is very incomplete.

> It is missing a background explanation on the algorithms, what they are for and their applications.

> Having only a list of names with some references is very limited and at least a paragraph explaining each item should be included.

Misc

>  In the paper the authors use "I" and "We".

> Just one should be used in all paper
