### 2023-08-08 #5664 @kokitsuyuzaki
Yes, I'm working on the comments.

This has been a difficult process due to many comments.
Since many of the comments from two reviewers are duplicated, I have combined them as Issues.
e.g.,
https://github.com/rikenbit/dcTensor/issues/8

I will reply to all of them in the next few days.
