### 2023-08-09 rikenbit/dcTensor#8 @kokitsuyuzaki
> Summary

Comments from @dekuenstle and @CeciliaCoelho are conflicting; @dekuenstle recommends adding text, while @CeciliaCoelho recommends shortening the summary, down to about one paragraph.

In my opinion, there is no need to shorten the summary; No such guidelines are written in the JOSS Documentation and, checking other Accepted JOSS papers, there are about 1 to 4 paragraphs depending on the complexity of the contents.

For that reason, I will follow @dekuenstle's opinion here.
