+++ paper/paper.md
@@ -19,30 +19,32 @@ bibliography: paper.bib
 ---
 
 # Summary
+Matrix factorization (MF) is a widely used approach to extract significant patterns in a data matrix. MF is formalized as the approximation of a data matrix $X$ by the matrix product of two factor matrices $U$ and $V$. Because this formalization has a large number of degrees of freedom, some constraints are imposed on the solution. Non-negative matrix factorization (NMF) imposing a non-negative solution for the factor matrices is a widely used algorithm to decompose non-negative matrix data matrix. Due to the interpretability of its non-negativity and the convenience of using decomposition results as clustering, there are many applications of NMF in image processing, audio processing, and bioinformatics [@amari].
 
-Non-negative matrix factorization (NMF) is a widely used algorithm to decompose non-negative matrix data into factor matrices. Due to the interpretability of its non-negativity and the convenience of using decomposition results as clustering, there are many applications of NMF in image processing, audio processing, and bioinformatics [@amari].
+A discrete version of NMF can also be considered by imposing a binary solution (e.g., {0,1}) for the factor matrices extracted from the data matrix and it is called binary matrix factorization (BMF) [@bmf]. BMF is recently featured in some data science domains such as market basket data, document-term data, Web click-stream data, DNA microarray expression profiles, or protein-protein complex interaction networks.
 
-We can consider the discrete version of NMF by imposing a binary solution ({0,1}) or ternary solution ({0,1,2}) for the factor matrices. Here, we will refer to these as binary matrix factorization (BMF) or ternary matrix factorization (TMF). It is also possible to apply such a constraint to only one of the two factor matrices, and we will call these (semi-) binary matrix factorization (SBMF) and (semi-) ternary matrix factorization (STMF).
-
-A BMF algorithm was first proposed by Zhang, Z. et al. in 2007 [@bmf], and the algorithm is based on a binary regularization against two factor matrices. Although their work focused on only BMF, the formulization can be applied to TMF, SBMF, or STMF. Besides, there is a growing demand to apply such (semi-) binary or (semi-) ternary decomposition to more heterogeneous non-negative data such as multiple matrices and tensors (high-dimensional arrays), which are higher-order data structures than matrices [@amari]. To meet these requirements, I originally developed \texttt{dcTensor}, which is an R/CRAN package to perform some discrete matrix/tensor decomposition algorithms (\url{https://cran.r-project.org/web/packages/dcTensor/index.html}).
[…]
