+++ paper/paper.bib
@@ -4,90 +4,109 @@
 
 % Binary NMF
 @article{bmf,
-  author = {Zhang, Z. et al.},
+  author = {{Zhang}, Z. and {Li}, T. and {Ding}, C. and {Zhang}, X.},
   journal = {ICDM 2007},
   title = {Binary Matrix Factorization with Applications},
   pages = {391-400},
+  doi = {10.1109/icdm.2007.99},
   year = {2007}}
+
+% Semi-Binary NMF
+@article{semi,
+  author = {{Ma}, X. and {Gao}, J. and {Liu}, X. and {Zhang}, T. and {Tang}, Y.},
+  journal = {MDPI mathematics},
+  title = {Probabilistic Non-Negative Matrix Factorization with Binary Components},
+  volumne = {9(11)},
+  pages = {1189},
+  doi = {10.3390/math9111189},
+  year = {2021}}
 
 % NMF, NTF, NTD
 @book{amari,
   title={Nonnegative Matrix and Tensor Factorizations},
-  author={CICHOCK, A. et al.},
+  author={{Cichocki}, A. and {Zdunek}, R. and {Phan}, A. H. and {Amari}, S.},
   year={2009},
   publisher={Wiley}
 }
 
 % NTF
 @article{ntf,
-  author = {Cichocki, A. et al.},
+  author={{Cichocki}, A. and {Zdunek}, R. and {Choi}, S. and {Plemmons}, R and {Amari}, S.},
   journal = {ICASSP '07},
   title = {Non-negative Tensor Factorization using Alpha and Beta Divergence},
   pages = {III-1393-III-1396},
+  doi = {10.1109/icassp.2007.367106},
   year = {2007}}
 
 % NTD
 @article{ntd,
-  author = {Kim, Y.-D. et al.},
+  author = {{Kim}, Y.-D. and {Choi}, S.},
   journal = {IEEE CVPR},
   title = {Nonnegative Tucker Decomposition},
   pages = {1-8},
+  doi = {10.1109/cvpr.2007.383405},
   year = {2007}}
 
 
 % NMF
 @article{nmf,
-  author = {Lee, D. and Seung, H.},
+  author = {{Lee}, D. and {Seung}, H.},
   journal = {Nature},
   title = {Learning the parts of objects by non-negative matrix factorization},
   volume = {401},
   pages = {788-791},
+  doi = {10.1038/44565},
   year = {1999}}
 
 % GD-SVD
 @article{svd,
-  author = {Tsuyuzaki, K. et al.},
+  author = {{Tsuyuzaki}, K. and {Sato}, H. and {Sato}, K. and {Nikaido}, I.},
   journal = {BMC Genome Biology},
   title = {Benchmarking principal component analysis for large-scale single-cell RNA-sequencing},
   volume = {21(1)},
   pages = {9},
+  doi = {10.1186/s13059-019-1900-3},
   year = {2020}}
 
 % siNMF
 @article{sinmf1,
-  author = {Badea, L.},
+  author = {{Badea}, L.},
   journal = {Pacific Symposium on Biocomputing},
   title = {Extracting Gene Expression Profiles Common to Colon and Pancreatic Adenocarcinoma using Simultaneous nonnegative matrix factorization},
   pages = {279-290},
+  doi = {10.1142/9789812776136_0027},
   year = {2008}}
 
 @article{sinmf2,
-  author = {Zhang, S. et al.},
+  author = {{Zhang}, S. {Liu}, C.-C. and {Li}, W. and {Shen}, H. and {Laird}, P. W. and {Zhou}, X. J.},
   journal = {Nucleic Acids Research},
   title = {Discovery of multi-dimensional modules by integrative analysis of cancer genomic data},
[…]
