### 2023-08-09 rikenbit/dcTensor#8 @kokitsuyuzaki
@dekuenstle 's comment

Summary

> The arguments could profit from a few more explanations.
> In the first paragraph, introducing matrix decomposition might be inclusive, i.e., explaining that the product of factor matrices reconstruct the original matrix and why this is a non-trivial problem.
> In the third paragraph, it remains to be seen why there is a demand for tensor decomposition.

I added the explanation of matrix decomposition in the first paragraph and the reasons for the extensions in the third paragraph.

Statement of Need

> However, this justification needs more explanation for which problems to solve and for which target community these implementations are needed.
> Especially it would be relevant to explain the need in applications for algorithms provided in the submitted package but not in related works.

I added such explanations in Statement of Need.

Related Work

> This also provides the opportunity to integrate the very short Related Work section into the Statement of Need.

I merged this section with Statement of Need.

Misc

> Overall, the language in the manuscript should be revised with a focus on rephrasing nested sentences and increasing coherence.

I used Grammarly and ChatGPT to improve the manuscript. If you still have any comments on my English, please be specific and say the line numbers.
