### 2023-08-09 rikenbit/dcTensor#8 @kokitsuyuzaki
@CeciliaCoelho 's comment

Summary

> The summary is too big in comparison to the rest of the documentation.
> The summary should describe concisely the package that is going to be presented stating what functions are implemented, some key features it contains and the target programming language.
> Right now the summary presented is more appropriate for the statement of need since it gives a background explanation of the implemented algorithms.
> The summary should be one paragraph. If the authors need some guidelines to fix the summary they should check accepted JOSS papers.

I cannot agree with the idea reducing the summary to one paragraph or moving some of it to the Statement of need.
Compared with accepted JOSS papers, I think the current sentence structure is not unnatural.
https://github.com/rikenbit/dcTensor/issues/8#issuecomment-1670492818

Statement of need

> The statement of need is very incomplete.
> It is missing a background explanation on the algorithms, what they are for and their applications.

I added some background explanations and their applications in Summary.

> Having only a list of names with some references is very limited and at least a paragraph explaining each item should be included.

I classified these items by nesting the list.
https://github.com/rikenbit/dcTensor/commit/d00293df9b86e2f7484f04ec822d22bb699f7bd4#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5

Misc

> In the paper the authors use "I" and "We".
> Just one should be used in all paper

I fixed some "We", but other "We" is correct, because it means "I and the readers" of the manuscript.
https://github.com/rikenbit/dcTensor/commit/d00293df9b86e2f7484f04ec822d22bb699f7bd4#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5
