+++ tests/testthat/test_mask_dNTF.R
@@ -1,48 +1,50 @@
-#
-# 3-order tensor
-#
-X <- dcTensor::toyModel("dNTF")
-M1 <- kFoldMaskTensor(X, seeds=12345)[[1]]
-M2 <- kFoldMaskTensor(X, seeds=54321)[[1]]
+test_that("dNTF (Mask)", {
+	#
+	# 3-order tensor
+	#
+	X <- dcTensor::toyModel("dNTF")
+	M1 <- kFoldMaskTensor(X, seeds=12345)[[1]]
+	M2 <- kFoldMaskTensor(X, seeds=54321)[[1]]
 
-out1_1 <- dNTF(X, M=M1, rank=3, algorithm="Frobenius", num.iter=1)
-out1_2 <- dNTF(X, M=M1, rank=3, algorithm="Frobenius", init="dNMF", num.iter=1)
-out1_3 <- dNTF(X, M=M1, rank=3, algorithm="Frobenius", init="Random", num.iter=1)
-out2 <- dNTF(X, M=M1, rank=3, algorithm="KL", num.iter=1)
-out3 <- dNTF(X, M=M1, rank=3, algorithm="IS", num.iter=1)
-out4 <- dNTF(X, M=M1, rank=3, algorithm="Beta", Beta=-1, num.iter=1)
+	out1_1 <- dNTF(X, M=M1, rank=3, algorithm="Frobenius", num.iter=1)
+	out1_2 <- dNTF(X, M=M1, rank=3, algorithm="Frobenius", init="dNMF", num.iter=1)
+	out1_3 <- dNTF(X, M=M1, rank=3, algorithm="Frobenius", init="Random", num.iter=1)
+	out2 <- dNTF(X, M=M1, rank=3, algorithm="KL", num.iter=1)
+	out3 <- dNTF(X, M=M1, rank=3, algorithm="IS", num.iter=1)
+	out4 <- dNTF(X, M=M1, rank=3, algorithm="Beta", Beta=-1, num.iter=1)
 
-expect_equivalent(length(out1_1), 8)
-expect_equivalent(length(out1_2), 8)
-expect_equivalent(length(out1_3), 8)
-expect_equivalent(length(out2), 8)
-expect_equivalent(length(out3), 8)
-expect_equivalent(length(out4), 8)
+	expect_equivalent(length(out1_1), 6)
+	expect_equivalent(length(out1_2), 6)
+	expect_equivalent(length(out1_3), 6)
+	expect_equivalent(length(out2), 6)
+	expect_equivalent(length(out3), 6)
+	expect_equivalent(length(out4), 6)
 
-out5_1 <- dNTF(X, M=M2, rank=3, algorithm="Frobenius", num.iter=1)
-out5_2 <- dNTF(X, M=M2, rank=3, algorithm="Frobenius", init="dNMF", num.iter=1)
-out5_3 <- dNTF(X, M=M2, rank=3, algorithm="Frobenius", init="Random", num.iter=1)
-out6 <- dNTF(X, M=M2, rank=3, algorithm="KL", num.iter=1)
-out7 <- dNTF(X, M=M2, rank=3, algorithm="IS", num.iter=1)
-out8 <- dNTF(X, M=M2, rank=3, algorithm="Beta", Beta=-1, num.iter=1)
+	out5_1 <- dNTF(X, M=M2, rank=3, algorithm="Frobenius", num.iter=1)
+	out5_2 <- dNTF(X, M=M2, rank=3, algorithm="Frobenius", init="dNMF", num.iter=1)
+	out5_3 <- dNTF(X, M=M2, rank=3, algorithm="Frobenius", init="Random", num.iter=1)
+	out6 <- dNTF(X, M=M2, rank=3, algorithm="KL", num.iter=1)
[…]
