### 2023-08-14 #5664 @kokitsuyuzaki
@dekuenstle @CeciliaCoelho

I'm sorry for taking time so long.
Now I have addressed all the issues, so please comment them.
https://github.com/rikenbit/dcTensor/issues
