+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: dcTensor
 Type: Package
 Title: Discrete Matrix/Tensor Decomposition
-Version: 1.2.2
+Version: 1.2.3
 Authors@R: c(person("Koki", "Tsuyuzaki", role = c("aut", "cre"),
                       email = "k.t.the-answer@hotmail.co.jp"))
 Depends:
