### 2023-08-23 #5664 @kokitsuyuzaki
Thanks, @diehlpk

I tagged the latest version as v1.2.3 and archieved by Zenodo.

The DOI is **10.5281/zenodo.8275544**

Please also check these pages.
https://github.com/rikenbit/dcTensor/releases/tag/v1.2.3
https://zenodo.org/record/8275544
