### 2023-08-23 #5664 @diehlpk
@kokitsuyuzaki it seems the Zenodo entry does not match with the paper.

```
dcTensor: An R package for discrete matrix/tensor decomposition

kokitsuyuzaki
```

Yous should use your name and not your GitHub handle.

Can you please fix that?

Note there is no need to generate a new DOI.
